--- rampup-backend/apps/student-crud/src/student-crud.controller.ts
@@ -1,12 +1,0 @@
-import { Controller, Get } from '@nestjs/common';
-import { StudentCrudService } from './student-crud.service';
-
-@Controller()
-export class StudentCrudController {
-  constructor(private readonly studentCrudService: StudentCrudService) {}
-
-  @Get()
-  getHello(): string {
-    return this.studentCrudService.getHello();
-  }
-}
